fix(GlobalExceptionHandler, MedicationServiceImpl, app-test.properties): add test properties, fix controller advice annotation and add medication not found exception in delete method

+++ src/main/java/com/sanitas/recuerdame/medications/services/MedicationServiceImpl.java
@@ -87,6 +87,9 @@ public class MedicationServiceImpl implements MedicationService{
 
     @Override
     public void deleteMedication(Long id) {
+        if (!medicationRepository.existsById(id)) {
+            throw new MedicationNotFoundException("Medication not found with id: " + id);
+        }
         medicationRepository.deleteById(id);
     }
 }
